requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\Anshuman\VisionSearchProject\app.py: 4
Flask == 2.0.1

# C:\Users\Anshuman\VisionSearchProject\handlers.py: 4
Pillow == 8.2.0

# C:\Users\Anshuman\VisionSearchProject\app.py: 2,3
Werkzeug == 2.0.1

# C:\Users\Anshuman\VisionSearchProject\functions.py: 11,12
# C:\Users\Anshuman\VisionSearchProject\rough.py: 10,11
# C:\Users\Anshuman\VisionSearchProject\search.py: 10,11
# C:\Users\Anshuman\VisionSearchProject\train.py: 10,11
clarifai_grpc == 7.4.0

# C:\Users\Anshuman\VisionSearchProject\functions.py: 3,4,5
# C:\Users\Anshuman\VisionSearchProject\rough.py: 3,4,5
# C:\Users\Anshuman\VisionSearchProject\search.py: 3,4,5
# C:\Users\Anshuman\VisionSearchProject\train.py: 3,4,5
cloudinary == 1.25.0

# C:\Users\Anshuman\VisionSearchProject\functions.py: 14,15
# C:\Users\Anshuman\VisionSearchProject\rough.py: 13,14
# C:\Users\Anshuman\VisionSearchProject\search.py: 14,15
# C:\Users\Anshuman\VisionSearchProject\train.py: 13,14
pymongo == 3.11.4

# C:\Users\Anshuman\VisionSearchProject\handlers.py: 1
requests == 2.25.1
gunicorn